more david review feedback
- move circuit evaluation to circuit module
- zstd compress test vectors
