+++ Cargo.toml
@@ -15,3 +15,4 @@ hex = { version = "0.4.3" }
 rand = "0.8.5"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = { version = "1.0" }
+zstd = "0.13.3"
